updated readme.md ands sh scripts

+++ stop.sh
@@ -1,4 +1,6 @@
 
-
 docker compose -f docker-compose-gen.yaml stop -t 5
-docker compose -f docker-compose-gen.yaml down
+
+if [ "$1" == "-k" ]; then
+  docker compose -f docker-compose-gen.yaml down
+fi
